+++ jwst/lib/set_telescope_pointing.py
@@ -13,7 +13,9 @@ from ..lib.engdb_tools import (
     ENGDB_BASE_URL,
     ENGDB_Service,
 )
+from .exposure_types import IMAGING_TYPES, FGS_GUIDE_EXP_TYPES
 
+TYPES_TO_UPDATE = IMAGING_TYPES
 
 # Setup logging
 logger = logging.getLogger(__name__)
@@ -41,16 +43,6 @@ SIFOV2V_DEFAULT = np.array(
      [-0.00226892608, 0., 0.99999742598]]
 )
 
-# JWST Exposures that are Fine Guidance exposures that actually
-# define the pointing.
-FGS_GUIDE_EXP_TYPES = [
-    'fgs_acq1',
-    'fgs_acq2',
-    'fgs_fineguide',
-    'fgs_id-image',
-    'fgs_id-stack',
-    'fgs_track',
-]
 
 # Degree, radian, angle transformations
 R2D = 180./np.pi
@@ -59,8 +51,14 @@ A2R = D2R/3600.
 R2A = 3600.*R2D
 
 # SIAF container
-SIAF = namedtuple("SIAF", ["v2ref", "v3ref", "v3idlyang", "vparity"])
-SIAF.__new__.__defaults__ = (None, None, None, None)
+# The names should correspond to the names in the ``wcsinfo`` schema.
+# It is populated by the SIAF values in the PRD database based
+# on APERNAME and UseAfterDate and used to populate the keywords
+# in Level1bModel data models.
+SIAF = namedtuple("SIAF", ["v2_ref", "v3_ref", "v3yangle", "vparity",
+                           "crpix1", "crpix2", "cdelt1", "cdelt2",
+                           "vertices_idl"])
+SIAF.__new__.__defaults__ = (None, ) * 9
 
 # Pointing container
 Pointing = namedtuple("Pointing", ["q", "j2fgs_matrix", "fsmcorr", "obstime"])
@@ -68,30 +66,28 @@ Pointing.__new__.__defaults__ = (None, None, None, None)
 
 # Transforms
 Transforms = namedtuple("Transforms",
-    [
-        'm_eci2j',            # ECI to J-Frame
-        'm_j2fgs1',           # J-Frame to FGS1
-        'm_sifov_fsm_delta',  # FSM correction
-        'm_fgs12sifov',       # FGS1 to SIFOV
-        'm_eci2sifov',        # ECI to SIFOV
-        'm_sifov2v',          # SIFOV to V1
-        'm_eci2v',            # ECI to V
-        'm_v2siaf',           # V to SIAF
-        'm_eci2siaf'          # ECI to SIAF
-    ])
+                        [
+                            'm_eci2j',            # ECI to J-Frame
+                            'm_j2fgs1',           # J-Frame to FGS1
+                            'm_sifov_fsm_delta',  # FSM correction
+                            'm_fgs12sifov',       # FGS1 to SIFOV
+                            'm_eci2sifov',        # ECI to SIFOV
+                            'm_sifov2v',          # SIFOV to V1
+                            'm_eci2v',            # ECI to V
+                            'm_v2siaf',           # V to SIAF
+                            'm_eci2siaf'          # ECI to SIAF
+                        ])
 Transforms.__new__.__defaults__ = (None, None, None, None, None,
-                                    None, None, None, None)
+                                   None, None, None, None)
 # WCS reference container
-WCSRef = namedtuple(
-    'WCSRef',
-    ['ra', 'dec', 'pa'])
+WCSRef = namedtuple( 'WCSRef', ['ra', 'dec', 'pa'])
 WCSRef.__new__.__defaults__ = (None, None, None)
 
 
 def add_wcs(filename, default_pa_v3=0., siaf_path=None, engdb_url=None,
             tolerance=60, allow_default=False, reduce_func=None,
             dry_run=False, **transform_kwargs):
-    """Add WCS information to a FITS file
+    """Add WCS information to a FITS file.
 
     Telescope orientation is attempted to be obtained from
     the engineering database. Failing that, a default pointing
@@ -102,7 +98,7 @@ def add_wcs(filename, default_pa_v3=0., siaf_path=None, engdb_url=None,
     Parameters
     ----------
     filename: str
-        The path to a data file
+        The path to a data file.
 
     default_pa_v3: float
         The V3 position angle to use if the pointing information
@@ -140,18 +136,38 @@ def add_wcs(filename, default_pa_v3=0., siaf_path=None, engdb_url=None,
     Currently it only uses a constant value for the engineering keywords
     since the Engineering Database does not yet contain them.
 
-    It assumes the following keywords are present in the file header:
+    It starts by populating the headers with values from the SIAF database.
+    It adds the following keywords to all files:
 
     V2_REF (arcseconds)
     V3_REF (arcseconds)
     VPARITY (+1 or -1)
     V3I_YANG (decimal degrees)
 
-    The keywords added are:
+    In addition for ``IMAGING_MODES`` it adds these keywords:
+
+    CRPIX1 (pixels)
+    CRPIX2 (pixels)
+    CDELT1 (deg/pix)
+    CDELT2 (deg/pix)
+    CUNIT1 (str)
+    CUNIT2
+    CTYPE1
+    CTYPE2
+    WCSAXES
+
+    The keywords computed and added to all files are:
 
     RA_V1
     DEC_V1
     PA_V3
+    RA_REF
+    DEC_REF
+    ROLL_REF
+    S_REGION
+
+    In addition the following keywords are computed and added to IMAGING_MODES only:
+
     CRVAL1
     CRVAL2
     PC1_1
@@ -174,10 +190,6 @@ def add_wcs(filename, default_pa_v3=0., siaf_path=None, engdb_url=None,
         reduce_func=reduce_func,
         **transform_kwargs
     )
-    try:
-        _add_axis_3(model)
-    except Exception:
-        pass
     model.meta.model_type = None
 
     if dry_run:
@@ -237,13 +249,23 @@ def update_wcs(model, default_pa_v3=0., siaf_path=None, engdb_url=None,
         exp_type = model.meta.exposure.type.lower()
     except AttributeError:
         exp_type = None
+    aperture_name = model.meta.aperture.name.upper()
+    if aperture_name != "UNKNOWN":
+        logger.info("Updating WCS for aperture {}".format(aperture_name))
+        useafter = model.meta.observation.date
+        siaf = _get_wcs_values_from_siaf(aperture_name, useafter, siaf_path)
+        populate_model_from_siaf(model, siaf)
+    else:
+        logger.warning("Aperture name is set to 'UNKNOWN'. "
+                       "WCS keywords will not be populated from SIAF.")
+
     if exp_type in FGS_GUIDE_EXP_TYPES:
         update_wcs_from_fgs_guiding(
             model, default_pa_v3=default_pa_v3
         )
     else:
         update_wcs_from_telem(
-            model, default_pa_v3=default_pa_v3, siaf_path=siaf_path, engdb_url=engdb_url,
+            model, default_pa_v3=default_pa_v3, siaf=siaf, engdb_url=engdb_url,
             tolerance=tolerance, allow_default=allow_default,
             reduce_func=reduce_func, **transform_kwargs
         )
@@ -305,9 +327,9 @@ def update_wcs_from_fgs_guiding(model, default_pa_v3=0.0, default_vparity=1):
 
 
 def update_wcs_from_telem(
-        model, default_pa_v3=0., siaf_path=None, engdb_url=None,
-        tolerance=0, allow_default=False,
-        reduce_func=None, **transform_kwargs
+    model, default_pa_v3=0., siaf=None, engdb_url=None,
+    tolerance=0, allow_default=False,
+    reduce_func=None, **transform_kwargs
 ):
     """Update WCS pointing information
 
@@ -353,19 +375,13 @@ def update_wcs_from_telem(
     # Get the SIAF and observation parameters
     obsstart = model.meta.exposure.start_time
     obsend = model.meta.exposure.end_time
-    siaf = SIAF(
-        v2ref=model.meta.wcsinfo.v2_ref,
-        v3ref=model.meta.wcsinfo.v3_ref,
-        v3idlyang=model.meta.wcsinfo.v3yangle,
-        vparity=model.meta.wcsinfo.vparity
-    )
     if None in siaf:
         if allow_default:
             logger.warning(
                 'Insufficient SIAF information found in header.'
                 'Resetting to a unity SIAF.'
             )
-            siaf = SIAF(0., 0., 0., 1)
+            siaf = SIAF(0, 0, 0, 1, 0, 0, 1, 1, [])
         else:
             raise ValueError('Insufficient SIAF information found in header.')
 
@@ -413,29 +429,25 @@ def update_wcs_from_telem(
 
     # Update Aperture pointing
     model.meta.aperture.position_angle = wcsinfo.pa
-    model.meta.wcsinfo.crval1 = wcsinfo.ra
-    model.meta.wcsinfo.crval2 = wcsinfo.dec
     model.meta.wcsinfo.ra_ref = wcsinfo.ra
     model.meta.wcsinfo.dec_ref = wcsinfo.dec
     model.meta.wcsinfo.roll_ref = compute_local_roll(
-        vinfo.pa, wcsinfo.ra, wcsinfo.dec, siaf.v2ref, siaf.v3ref
+        vinfo.pa, wcsinfo.ra, wcsinfo.dec, siaf.v2_ref, siaf.v3_ref
     )
-    (model.meta.wcsinfo.pc1_1,
-     model.meta.wcsinfo.pc1_2,
-     model.meta.wcsinfo.pc2_1,
-     model.meta.wcsinfo.pc2_2) = calc_rotation_matrix(
-         wcsinfo.pa * D2R, vparity=siaf.vparity
-     )
-
-    wcsaxes = model.meta.wcsinfo.wcsaxes
-    if wcsaxes is None:
-        wcsaxes = 2
-    model.meta.wcsinfo.wcsaxes = max(2, wcsaxes)
+    if model.meta.exposure.type.lower() in TYPES_TO_UPDATE:
+        model.meta.wcsinfo.crval1 = wcsinfo.ra
+        model.meta.wcsinfo.crval2 = wcsinfo.dec
+        (model.meta.wcsinfo.pc1_1,
+         model.meta.wcsinfo.pc1_2,
+         model.meta.wcsinfo.pc2_1,
+         model.meta.wcsinfo.pc2_2) = calc_rotation_matrix(
+             wcsinfo.pa * D2R, vparity=siaf.vparity
+         )
 
     # Calculate S_REGION with the footprint
     # information
     try:
-        update_s_region(model, prd_db_filepath=siaf_path)
+        update_s_region(model, siaf)
     except Exception as e:
         logger.warning(
             'Calculation of S_REGION failed and will be skipped.'
@@ -443,7 +455,7 @@ def update_wcs_from_telem(
         )
 
 
-def update_s_region(model, prd_db_filepath=None):
+def update_s_region(model, siaf):
     """Update ``S_REGION`` sky footprint information.
 
     The ``S_REGION`` keyword is intended to store the spatial footprint of
@@ -453,29 +465,14 @@ def update_s_region(model, prd_db_filepath=None):
     ----------
     model : `~jwst.datamodels.DataModel`
         The model to update in-place.
-    prd_db_filepath : str
-        The filepath to the SIAF file (PRD database).
-        If None, attempt to get the path from the ``XML_DATA`` environment variable.
+    siaf : namedtuple
+        The ``SIAF`` tuple withg values populated from the PRD database.
     """
-    if prd_db_filepath is None:
-        try:
-            prd_db_filepath = os.path.join(os.environ['XML_DATA'], "prd.db")
-        except KeyError:
-            logger.info("Unknown path to PRD DB file: {0}".format(prd_db_filepath))
-            return
-    if not os.path.exists(prd_db_filepath):
-        logger.info("Invalid path to PRD DB file: {0}".format(prd_db_filepath))
-        return
-    aperture_name = model.meta.aperture.name
-    useafter = model.meta.observation.date
-    vertices = _get_vertices_idl(
-        aperture_name, useafter, prd_db_filepath=prd_db_filepath
-    )
-    vertices = list(vertices.values())[0]
+    vertices = siaf.vertices_idl
     xvert = vertices[:4]
     yvert = vertices[4:]
     logger.info(
-        "Vertices for aperture {0}: {1}".format(aperture_name, vertices)
+        "Vertices for aperture {0}: {1}".format(model.meta.aperture.name, vertices)
     )
     # Execute IdealToV2V3, followed by V23ToSky
     from ..transforms.models import IdealToV2V3, V23ToSky
@@ -741,7 +738,7 @@ def calc_aperture_wcs(m_eci2siaf):
     # The VyPA @ xref,yref is given by
     y = cos(vy_dec) * sin(vy_ra-wcs_ra)
     x = sin(vy_dec) * cos(wcs_dec) - \
-        cos(vy_dec) * sin(wcs_dec) * cos((vy_ra - wcs_ra))
+      cos(vy_dec) * sin(wcs_dec) * cos((vy_ra - wcs_ra))
     wcs_pa = np.arctan2(y, x)
 
     # Convert all WCS to degrees
@@ -934,7 +931,8 @@ def calc_v2siaf_matrix(siaf):
     transform: np.array((3, 3))
         The V1 to SIAF transformation matrix
     """
-    v2, v3, v3idlyang, vparity = siaf
+    v2, v3, v3idlyang, vparity = (siaf.v2_ref, siaf.v3_ref,
+                                  siaf.v3yangle, siaf.vparity)
     v2 *= A2R
     v3 *= A2R
     v3idlyang *= D2R
@@ -975,7 +973,7 @@ def calc_position_angle(v1, v3):
     """
     y = cos(v3.dec) * sin(v3.ra-v1.ra)
     x = sin(v3.dec) * cos(v1.dec) - \
-        cos(v3.dec) * sin(v1.dec) * cos((v3.ra - v1.ra))
+      cos(v3.dec) * sin(v1.dec) * cos((v3.ra - v1.ra))
     v3_pa = np.arctan2(y, x)
 
     return v3_pa
@@ -1110,54 +1108,91 @@ def _roll_angle_from_matrix(matrix, v2, v3):
     return new_roll
 
 
-def _get_vertices_idl(aperture_name, useafter, prd_db_filepath=None):
+def _get_wcs_values_from_siaf(aperture_name, useafter, prd_db_filepath=None):
+    """
+    Query the SIAF database file and get WCS values.
+
+    Given an ``APERTURE_NAME`` and a ``USEAFTER`` date query the SIAF database
+    and extract the following keywords:
+    ``V2Ref``, ``V3Ref``, ``V3IdlYAngle``, ``VIdlParity``,
+    ``XSciRef``, ``YSciRef``, ``XSciScale``, ``YSciScale``,
+    ``XIdlVert1``, ``XIdlVert2``, ``XIdlVert3``, ``XIdlVert4``,
+    ``YIdlVert1``, ``YIdlVert2``, ``YIdlVert3``, ``YIdlVert4``
+
+    Parameters
+    ----------
+    aperture_name : str
+        The name of the aperture in the data file (``model.meta.aperture.name``).
+    useafter : str
+        The date of observation (``model.meta.date``)
+    prd_db_filepath : str
+        The path to the SIAF (PRD database) file.
+        If None, attempt to get the path from the ``XML_DATA`` environment variable.
+
+    Returns
+    -------
+    siaf : namedtuple
+        The SIAF namedtuple with values from the PRD database.
+    """
+    if prd_db_filepath is None:
+        try:
+            prd_db_filepath = os.path.join(os.environ['XML_DATA'], "prd.db")
+        except KeyError:
+            message = "Unknown path to PRD DB file or missing env variable ``XML_DATA``."
+            log.info(message)
+            raise KeyError(message)
+    if not os.path.exists(prd_db_filepath):
+        message = "Invalid path to PRD DB file: {0}".format(prd_db_filepath)
+        log.info(message)
+        raise OSError(message)
     prd_db_filepath = "file:{0}?mode=ro".format(prd_db_filepath)
     logger.info("Using SIAF database from {}".format(prd_db_filepath))
-    logger.info("Getting aperture vertices for aperture "
-             "{0} with USEAFTER {1}".format(aperture_name, useafter))
+    logger.info("Quering SIAF for aperture "
+                "{0} with USEAFTER {1}".format(aperture_name, useafter))
     aperture = (aperture_name, useafter)
 
     RESULT = {}
+    PRD_DB = False
     try:
         PRD_DB = sqlite3.connect(prd_db_filepath, uri=True)
 
         cursor = PRD_DB.cursor()
-        cursor.execute("SELECT Apername, XIdlVert1, XIdlVert2, XIdlVert3, XIdlVert4, "
+        cursor.execute("SELECT Apername, V2Ref, V3Ref, V3IdlYAngle, VIdlParity, "
+                       "XSciRef, YSciRef, XSciScale, YSciScale, "
+                       "XIdlVert1, XIdlVert2, XIdlVert3, XIdlVert4, "
                        "YIdlVert1, YIdlVert2, YIdlVert3, YIdlVert4 "
-                       "FROM Aperture WHERE Apername = ? and UseAfterDate <= ? ORDER BY UseAfterDate LIMIT 1", aperture)
+                       "FROM Aperture WHERE Apername = ? and UseAfterDate <= ? ORDER BY UseAfterDate LIMIT 1",
+                       aperture)
         for row in cursor:
-            RESULT[row[0]] = tuple(row[1:9])
+            RESULT[row[0]] = tuple(row[1:17])
         PRD_DB.commit()
-    except sqlite3.Error as err:
-        print("Error" + err.args[0])
+    except (sqlite3.Error, sqlite3.OperationalError) as err:
+        print("Error: " + err.args[0])
         raise
     finally:
         if PRD_DB:
             PRD_DB.close()
     logger.info("loaded {0} table rows from {1}".format(len(RESULT), prd_db_filepath))
-    return RESULT
-
-
-def _add_axis_3(model):
-    """
-    Adds CTYPE3 and CUNIT3 for spectral observations.
-    This is temporary, may be moved to SDP proccessing later.
-    """
-    wcsaxes = model.meta.wcsinfo.wcsaxes
-    if wcsaxes is not None and wcsaxes == 3:
-        model.meta.wcsinfo.ctype3 = "WAVE"
-        model.meta.wcsinfo.cunit3 = "um"
+    if RESULT:
+        values = list(RESULT.values())[0]
+        # This call populates the SIAF tuple with the values from the database.
+        # The last 8 values returned from the database are the vertices.
+        # They are wrapped in a list and assigned to SIAF.vertices_idl.
+        siaf = SIAF(*values[:-8], values[-8:])
+        return siaf
+    else:
+        return SIAF()
 
 
 def calc_rotation_matrix(angle, vparity=1):
-    """ Calculate the rotation matrix
+    """ Calculate the rotation matrix.
 
     Parameters
     ----------
-    angle: float in radians
-        The angle to create the matrix
+    angle : float in radians
+        The angle to create the matrix.
 
-    vparity: int
+    vparity : int
         The x-axis parity, usually taken from
         the JWST SIAF parameter VIdlParity.
         Value should be "1" or "-1".
@@ -1362,6 +1397,36 @@ def all_pointings(mnemonics):
         raise ValueError('No non-zero quanternion found.')
 
     return pointings
+
+
+
+def populate_model_from_siaf(model, siaf):
+    """
+    Populate the WCS keywords of a Level1bModel from the SIAF.
+
+    Parameters
+    ----------
+    model : `~jwst.datamodels.Level1bModel`
+        Input data as Level1bModel.
+    siaf : namedtuple
+        The WCS keywords read in from the SIAF.
+    """
+    # If not an imaging mode, update the pointing only.
+    model.meta.wcsinfo.v2_ref = siaf.v2_ref
+    model.meta.wcsinfo.v3_ref = siaf.v3_ref
+    model.meta.wcsinfo.v3yangle = siaf.v3yangle
+    model.meta.wcsinfo.vparity = siaf.vparity
+    if model.meta.exposure.type.lower() in TYPES_TO_UPDATE:
+        # For imaging modes update the pointing and
+        # the FITS WCS keywords.
+        model.meta.wcsinfo.ctype1 = 'RA--TAN'
+        model.meta.wcsinfo.ctype2 = 'DEC-TAN'
+        model.meta.wcsinfo.wcsaxes = 2
+        model.meta.wcsinfo.cunit1 = "deg"
+        model.meta.wcsinfo.cunit2 = "deg"
+        model.meta.wcsinfo.crpix1 = siaf.crpix1
+        model.meta.wcsinfo.crpix2 = siaf.crpix2
+        model.meta.coordinates.reference_frame = "ICRS"
 
 
 def first_pointing(mnemonics):