+++ CHANGES.rst
@@ -20,6 +20,8 @@ assign_wcs
 
  - The bounding box of the MIRI LRS WCS is now in "image" coordinates, not full frame. [#3063]
 
+ - FITS WCS keywords are written out only if the observation is one of the IMAGING_MODES. [#3066]
+
 associations
 ------------
 
@@ -279,6 +281,9 @@ scripts
 - ``set_telescope_pointing.py``: Update method of choosing pointing parameters. [#2900, #3008, #3022]
 
 - ``set_telescope_pointing.py``: Allow undefined SIAF. [#3002, #3006]
+
+- ``set_telescope_pointing`` now populates WCS keywords from the SIAF file. [#3066]
+
 
 0.12.2 (2018-11-15)
 ===================